Lambda/requirements_lambda.txt
boto3
requests>=2.28.0
python-dotenv==0.21.1
SQLAlchemy==1.4.46
PyMySQL==1.0.2
yfinance>=0.2.65
pandas==1.5.3
numpy==1.26.4